Move data structures out of `polkadot_node_subsystem_types` module up
