+++ polkadot/node/overseer/src/lib.rs
@@ -60,7 +60,6 @@
 #![allow(clippy::all)]
 
 pub mod collation_generation;
-mod polkadot_node_subsystem_types;
 mod polkadot_overseer_gen;
 
 use std::{
@@ -73,6 +72,8 @@ use std::{
 
 use futures::{future::FusedFuture, select, stream::FusedStream, FutureExt, StreamExt};
 use lru::LruCache;
+use smallvec::SmallVec;
+use std::fmt;
 
 use sc_client_api::{
 	BlockBackend, BlockImportNotification, BlockchainEvents, FinalityNotification,
@@ -81,10 +82,6 @@ use sp_api::ProvideRuntimeApi;
 use sp_blockchain::HeaderBackend;
 
 use crate::collation_generation::CollationGenerationMessage;
-pub use polkadot_node_subsystem_types::errors::SubsystemError;
-use polkadot_node_subsystem_types::{
-	errors::SubsystemResult, ActivatedLeaf, ActiveLeavesUpdate, LeafStatus, OverseerSignal,
-};
 
 use cirrus_node_primitives::{CollationGenerationConfig, ExecutorSlotInfo};
 use sp_executor::{BundleEquivocationProof, ExecutorApi, FraudProof, InvalidTransactionProof};
@@ -93,6 +90,136 @@ use subspace_runtime_primitives::{opaque::Block, BlockNumber, Hash};
 use polkadot_overseer_gen::{
 	FromOverseer, MessagePacket, SignalsReceived, SubsystemInstance, TimeoutExt,
 };
+
+/// How many slots are stack-reserved for active leaves updates
+///
+/// If there are fewer than this number of slots, then we've wasted some stack space.
+/// If there are greater than this number of slots, then we fall back to a heap vector.
+const ACTIVE_LEAVES_SMALLVEC_CAPACITY: usize = 8;
+
+/// The status of an activated leaf.
+#[derive(Debug, Clone)]
+pub enum LeafStatus {
+	/// A leaf is fresh when it's the first time the leaf has been encountered.
+	/// Most leaves should be fresh.
+	Fresh,
+	/// A leaf is stale when it's encountered for a subsequent time. This will happen
+	/// when the chain is reverted or the fork-choice rule abandons some chain.
+	Stale,
+}
+
+impl LeafStatus {
+	/// Returns a `bool` indicating fresh status.
+	pub fn is_fresh(&self) -> bool {
+		match *self {
+			LeafStatus::Fresh => true,
+			LeafStatus::Stale => false,
+		}
+	}
+
+	/// Returns a `bool` indicating stale status.
+	pub fn is_stale(&self) -> bool {
+		match *self {
+			LeafStatus::Fresh => false,
+			LeafStatus::Stale => true,
+		}
+	}
+}
+
+/// Activated leaf.
+#[derive(Debug, Clone)]
+pub struct ActivatedLeaf {
+	/// The block hash.
+	pub hash: Hash,
+	/// The block number.
+	pub number: BlockNumber,
+	/// The status of the leaf.
+	pub status: LeafStatus,
+}
+
+/// Changes in the set of active leaves: the parachain heads which we care to work on.
+///
+/// Note that the activated and deactivated fields indicate deltas, not complete sets.
+#[derive(Clone, Default)]
+pub struct ActiveLeavesUpdate {
+	/// New relay chain block of interest.
+	pub activated: Option<ActivatedLeaf>,
+	/// Relay chain block hashes no longer of interest.
+	pub deactivated: SmallVec<[Hash; ACTIVE_LEAVES_SMALLVEC_CAPACITY]>,
+}
+
+impl ActiveLeavesUpdate {
+	/// Create a `ActiveLeavesUpdate` with a single activated hash
+	pub fn start_work(activated: ActivatedLeaf) -> Self {
+		Self { activated: Some(activated), ..Default::default() }
+	}
+
+	/// Create a `ActiveLeavesUpdate` with a single deactivated hash
+	pub fn stop_work(hash: Hash) -> Self {
+		Self { deactivated: [hash][..].into(), ..Default::default() }
+	}
+
+	/// Is this update empty and doesn't contain any information?
+	pub fn is_empty(&self) -> bool {
+		self.activated.is_none() && self.deactivated.is_empty()
+	}
+}
+
+impl PartialEq for ActiveLeavesUpdate {
+	/// Equality for `ActiveLeavesUpdate` doesn't imply bitwise equality.
+	///
+	/// Instead, it means equality when `activated` and `deactivated` are considered as sets.
+	fn eq(&self, other: &Self) -> bool {
+		self.activated.as_ref().map(|a| a.hash) == other.activated.as_ref().map(|a| a.hash) &&
+			self.deactivated.len() == other.deactivated.len() &&
+			self.deactivated.iter().all(|a| other.deactivated.contains(a))
+	}
+}
+
+impl fmt::Debug for ActiveLeavesUpdate {
+	fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+		f.debug_struct("ActiveLeavesUpdate")
+			.field("activated", &self.activated)
+			.field("deactivated", &self.deactivated)
+			.finish()
+	}
+}
+
+/// Signals sent by an overseer to a subsystem.
+#[derive(PartialEq, Clone, Debug)]
+pub enum OverseerSignal {
+	/// Subsystems should adjust their jobs to start and stop work on appropriate block hashes.
+	ActiveLeaves(ActiveLeavesUpdate),
+	/// `Subsystem` is informed of a new slot.
+	NewSlot(ExecutorSlotInfo),
+	/// Conclude the work of the `Overseer` and all `Subsystem`s.
+	Conclude,
+}
+
+/// An error type that describes faults that may happen
+///
+/// These are:
+///   * Channels being closed
+///   * Subsystems dying when they are not expected to
+///   * Subsystems not dying when they are told to die
+///   * etc.
+#[derive(thiserror::Error, Debug)]
+#[allow(missing_docs)]
+pub enum SubsystemError {
+	#[error(transparent)]
+	QueueError(#[from] futures::channel::mpsc::SendError),
+
+	/// Generated by the `#[overseer(..)]` proc-macro
+	#[error(transparent)]
+	Generated(#[from] crate::polkadot_overseer_gen::OverseerError),
+
+	/// Generated by the `#[overseer(..)]` proc-macro
+	#[error(transparent)]
+	RuntimeApi(#[from] sp_api::ApiError),
+}
+
+/// Ease the use of subsystem errors.
+type SubsystemResult<T> = Result<T, self::SubsystemError>;
 
 /// Store 2 days worth of blocks, not accounting for forks,
 /// in the LRU cache. Assumes a 6-second block time.
